[IMP] l10n_lu: 2025 coa and vat report updates
Description of the issue this commit addresses:

In january 2025, updates were made to some sections of the LU tax report and
chart of accounts. Therefore, the version we use is out of date.

Desired behavior after this commit is merged:

The COA and VAT reports use the new values updated in 2025.

task-4587067

closes odoo/odoo#204858

X-original-commit: 8515b7751ed8767709d191bc31d2c0247ad8c16a
Related: odoo/enterprise#82912
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>