[IMP] mail, *: define JS model fields with 'fields.'
Before this commit, JS model field definition was made with
following syntax:

```js
class Thread extends Record {
	welcomeMessage = Record.one("Message");
	messages = Record.many("Message");
	last_interest_dt = Record.datetime();
	name = Record.attr();
}
```

They use `Record.` static methods to define fields.
The good aspect of it is that features are all available in `Record`.

In an effort to simplify JS model and harmonize with python ORM,
this commit change the syntax of defining field as follow:

```js
class Thread extends Record {
	welcomeMessage = fields.One("Message");
	messages = fields.Many("Message");
	last_interest_dt = fields.Datetime();
	name = fields.Attr();
}
```

This is a small step in matching the syntax with python ORM.
Another benefit of this new syntax, is that this doesn't rely on
static methods, therefore the autocompletion / docstring works
much better.

closes odoo/odoo#204634

Related: odoo/enterprise#82811
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>